# server/internal/store/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package store

import (
	"time"

	"github.com/google/uuid"
)

type Connection struct {
	ID                 uuid.UUID  `json:"id"`
	UserID             uuid.UUID  `json:"user_id"`
	Platform           string     `json:"platform"`
	ExternalAccountID  string     `json:"external_account_id"`
	Username           string     `json:"username"`
	AccessTokenEnc     string     `json:"access_token_enc"`
	TokenExpiresAt     *time.Time `json:"token_expires_at"`
	Scopes             []string   `json:"scopes"`
	Status             string     `json:"status"`
	SubscriptionStatus string     `json:"subscription_status"`
	SubscribedFields   []string   `json:"subscribed_fields"`
	LastRefreshedAt    *time.Time `json:"last_refreshed_at"`
	CreatedAt          time.Time  `json:"created_at"`
	UpdatedAt          time.Time  `json:"updated_at"`
}

type Rule struct {
	ID              uuid.UUID `json:"id"`
	ConnectionID    uuid.UUID `json:"connection_id"`
	MediaID         *string   `json:"media_id"`
	Keywords        []string  `json:"keywords"`
	MatchMode       string    `json:"match_mode"`
	ResponseMessage string    `json:"response_message"`
	ResponseLink    *string   `json:"response_link"`
	RequireFollow   bool      `json:"require_follow"`
	CaptureEmail    bool      `json:"capture_email"`
	Status          string    `json:"status"`
	CreatedAt       time.Time `json:"created_at"`
	UpdatedAt       time.Time `json:"updated_at"`
}

type User struct {
	ID          uuid.UUID  `json:"id"`
	LastLoginAt *time.Time `json:"last_login_at"`
	CreatedAt   time.Time  `json:"created_at"`
	UpdatedAt   time.Time  `json:"updated_at"`
}